README Update and Creation of local Scripts
I've updated the README with a more thorough description of the pipeline workflow to hopefully clarify what one might need to do to get this code running on their own system or a different cluster.

Began working on a python script to be able to run a fit on a local system; this will take a bit to write because quite a lot of the code makes the assumption it's being run on the SLURM-based MSI system. Ideally this code will be agnostic enough that it can serve as the example script for others to run when first using the library.

While not in the actual commit, I've made a handful of issues that should be addressed in the future. One important task is to decouple the code from reliance on MSI and make it easier for others to get the code running on an agnostic system. It would also be helpful to organize the repo in a way that makes it easier to browse without breaking the msi automation.