abwysqy variants
Created a .csv file for the abwysqy data from (Tomas' paper)[https://ui.adsabs.harvard.edu/abs/2021NatAs...5..917A/abstract] as well as variants of the original abwysqy data that was stalling when being run on the MSI pipeline. To investigate this, I created a handful of variants that reduce the number of non-detections in the file:

- Shortened1: All non-detections are removed
- Shortened2: Only non-detections from prior to the first detection are removed
- Shortened3: Non-detections more than one day before the first detection are removed
- Shortened4: The only i-filter data point, a non-detection, is removed

I'll add these to be processed by the MSI pipeline and investigate which are still stalling

I also added the manual_plotter.py file for posterity, even though it's not particularly useful outside the very specific case in which it was put together. Most of it is based on existing code and flags are hard coded to work with a specific file that needed to be fit.