Job Timing
Changed the assigned length of jobs to be just under 8 hours compared to the previous limit of 12 hours. Also adjusted the timeout check in make_jobs to default to 8 hours. In addition, on line 188 of make_jobs.py, I cast the argument as an int explicitly to make sure the conditional is actually be evaluated correctly. Double check if this breaks behavior