v2 of paper fits (failed)
Added the attempted 2nd automatic fitting of the paper candidates, using the modified files that were created for abwysqy. Seems to be possibly a coinciding issue of both MSI and something with the model fitting. The log file may be clarifying in this case, but note that it has been zipped to comply with Github file limitations. Most of the log file is very repetitive, so it went from 150 MB to ~40 MB. May end up having to just run the fits manually on local system