localScript prototyping + prior relocation
Wrote what seems to be a functional version of a local script to run fits, which uses the --plot flag in light_curve_analysis to (hopefully) plot the best fit lightcurve using nmma rather than the code block present in nmma_fit. If this works, could be applied to nmma_fit.py, shaving down the size of the script by a bit over a third.

Relocated the priors into their own folders and tweaked nmma_fit.py so it isn't broken by the new prior paths.

Started a small python script called fetch_priors.py. While not currently functional, hopefully we could eventually make it something people can dynamically add new priors/models to so we don't have to go in and tweak all the settings to allow for new models. This would require reworking of several of the other scripts to make them model agnostic, or rather to pull out any model names/directories into arguments that can be provided to those scripts by other scripts/functions in the pipeline. Would have to discuss with Michael/nmma group how we might want to approach adding new models to the main codebase so as to preserve functionality (I would assume pull request that only adds a couple of lines to one or two scripts that contain information on priors and models)