naming convention tweak
Did a small change to the way the plots and logs are named in both localScript.py and nmma_fit.py. Namely, I altered it so the format is now "[candidate]_[model]]" as opposed to the old format of "[model]". Should be helpful for sharing plots and not confusing what's what

Hopefully it doesn't break fit_bot.py on schoty, but it might! If that happens, look around line 270 of fit_bot.py on schoty and try to tweak it so it stops being broken :)