stats plotting start
I've written a few basic functions that will make plots for some of the stats of the project (it should probably function, but haven't tested on MSI yet). Currently handles the stats for some of the candidate stuff, but none for the fit stats yet. Comments explain further in stats.py.

With the way stats.py is written, it's starting to look at least something like a python package structure, but I'd need to work on that, probably post paper.