Update stats.py
moved around functions in stats.py to better organize it. also created get_json, which is the better version of get_sampling_time(). While not present in this commit, I've written out a way to generate a pandas dataframe that can be used to more easily get all of the statistics desired for these plots. This will require tweaking the plots but should overall be a better implementation over the current series of lists