get_dataframe() creation
Wrote out get_dataframe(), which should allow for the creation/reading of a dataframe that encodes most of the information needed to create our desired plots by using the df.loc[] method or similar to maintain performance. Further info is provided in the stats.py file, as the function should be fairly well documented.

Note: Copilot is kind of scary with how useful it is