Update stats.py
completed the plotUnfit() function, includes a collection of 8 plots and histograms that provide some visualizations on items that weren't fit.

This function was a lot easier to write on account of the dataframe, so modifying the other plotting functions should probably be easy to do. Haven't tested how performant the dataframe is for large data collections, though I think we've had <5000 fits done in total up to this point